refactor(logic): 分離されたロジック型を新しいファイルに移動し、整理する
