+++ package/main/src/types/logic/and.ts
@@ -1,0 +1,7 @@
+import type { isBoolean } from "./isBoolean";
+
+export type AND<X, Y> = isBoolean<X> extends true
+  ? isBoolean<Y> extends true
+    ? true
+    : false
+  : false;
