+++ package/main/src/types/logic/capitalize.ts
@@ -1,0 +1,4 @@
+// 文字の先頭を大文字にする型
+export type Capitalize<S extends string> = S extends `${infer F}${infer R}`
+  ? `${Uppercase<F>}${R}`
+  : S;
