+++ package/main/src/types/logic/first.ts
@@ -1,0 +1,4 @@
+// 配列の先頭を取得する型
+export type First<T extends unknown[]> = T extends [infer F, ...unknown[]]
+  ? F
+  : never;
