+++ package/main/src/types/logic/imply.ts
@@ -1,0 +1,4 @@
+import type { isBoolean } from "./isBoolean";
+import type { NOT } from "./not";
+
+export type IMPLY<X, Y> = NOT<X> extends true ? true : isBoolean<Y>;
