+++ package/main/src/types/logic/length.ts
@@ -1,0 +1,6 @@
+// 配列の長さを取得する型
+export type Length<T extends unknown[]> = T extends { length: infer L }
+  ? L extends number
+    ? L
+    : never
+  : never;
