+++ package/main/src/types/logic/lengthOfString.ts
@@ -1,0 +1,4 @@
+import type { StringToArray } from "./stringToArray";
+
+// 文字を配列に変換して長さを取得する型
+export type LengthOfString<S extends string> = StringToArray<S>["length"];
