+++ package/main/src/types/logic/nand.ts
@@ -1,0 +1,4 @@
+import type { AND } from "./and";
+import type { NOT } from "./not";
+
+export type NAND<X, Y> = NOT<AND<X, Y>>;
