+++ package/main/src/types/logic/nor.ts
@@ -1,0 +1,4 @@
+import type { NOT } from "./not";
+import type { OR } from "./or";
+
+export type NOR<X, Y> = NOT<OR<X, Y>>;
