+++ package/main/src/types/logic/not.ts
@@ -1,0 +1,3 @@
+import type { isBoolean } from "./isBoolean";
+
+export type NOT<X> = isBoolean<X> extends true ? false : true;
