+++ package/main/src/types/logic/or.ts
@@ -1,0 +1,7 @@
+import type { isBoolean } from "./isBoolean";
+
+export type OR<X, Y> = isBoolean<X> extends true
+  ? true
+  : isBoolean<Y> extends true
+    ? true
+    : false;
