+++ package/main/src/types/logic/pop.ts
@@ -1,0 +1,4 @@
+// 配列の末尾を削除する型
+export type Pop<T extends unknown[]> = T extends [...infer R, unknown]
+  ? R
+  : never;
