+++ package/main/src/types/logic/popString.ts
@@ -1,0 +1,4 @@
+// 文字列の末尾を削除する型
+export type PopString<S extends string> = S extends `${infer R}${string}`
+  ? R
+  : never;
