+++ package/main/src/types/logic/shiftString.ts
@@ -1,0 +1,4 @@
+// 文字列の先頭を削除する型
+export type ShiftString<S extends string> = S extends `${string}${infer R}`
+  ? R
+  : never;
