+++ package/main/src/types/logic/xnor.ts
@@ -1,0 +1,4 @@
+import type { NOT } from "./not";
+import type { XOR } from "./xor";
+
+export type XNOR<X, Y> = NOT<XOR<X, Y>>;
