+++ package/main/src/types/logicType.ts
@@ -1,100 +1,8 @@
+import type { Length } from "./logic/length";
+import type { LengthOfString } from "./logic/lengthOfString";
+import type { Shift } from "./logic/shift";
+import type { ShiftString } from "./logic/shiftString";
 import type { Subtract } from "./math/subtract";
-
-// booleanに変換する型
-export type isBoolean<X> = X extends number
-  ? X extends 0
-    ? false
-    : true
-  : X extends string
-    ? X extends ""
-      ? false
-      : true
-    : X extends boolean
-      ? X
-      : X extends undefined
-        ? false
-        : X extends null
-          ? false
-          : X extends object
-            ? X extends unknown[]
-              ? X extends []
-                ? false
-                : true
-              : true
-            : X extends (...arguments_: unknown[]) => unknown
-              ? false
-              : true;
-
-export type AND<X, Y> = isBoolean<X> extends true
-  ? isBoolean<Y> extends true
-    ? true
-    : false
-  : false;
-
-export type OR<X, Y> = isBoolean<X> extends true
-  ? true
-  : isBoolean<Y> extends true
-    ? true
-    : false;
-
-export type XOR<X, Y> = isBoolean<X> extends true
-  ? isBoolean<Y> extends true
-    ? false
-    : true
-  : isBoolean<Y> extends true
-    ? true
-    : false;
-export type NOT<X> = isBoolean<X> extends true ? false : true;
-export type NAND<X, Y> = NOT<AND<X, Y>>;
-export type NOR<X, Y> = NOT<OR<X, Y>>;
-export type XNOR<X, Y> = NOT<XOR<X, Y>>;
-export type IMPLY<X, Y> = NOT<X> extends true ? true : isBoolean<Y>;
-
-// 文字列を配列に変換する型
-export type StringToArray<
-  S extends string,
-  T extends unknown[] = [],
-> = S extends `${infer F}${infer R}` ? StringToArray<R, [...T, F]> : T;
-
-// 配列の長さを取得する型
-export type Length<T extends unknown[]> = T extends { length: infer L }
-  ? L extends number
-    ? L
-    : never
-  : never;
-
-// 文字を配列に変換して長さを取得する型
-export type LengthOfString<S extends string> = StringToArray<S>["length"];
-
-// 文字の先頭を大文字にする型
-export type Capitalize<S extends string> = S extends `${infer F}${infer R}`
-  ? `${Uppercase<F>}${R}`
-  : S;
-
-// 配列の先頭を削除する型
-export type Shift<T extends unknown[]> = T extends [unknown, ...infer R]
-  ? R
-  : never;
-
-// 配列の末尾を削除する型
-export type Pop<T extends unknown[]> = T extends [...infer R, unknown]
-  ? R
-  : never;
-
-// 配列の先頭を取得する型
-export type First<T extends unknown[]> = T extends [infer F, ...unknown[]]
-  ? F
-  : never;
-
-// 文字列の先頭を削除する型
-export type ShiftString<S extends string> = S extends `${string}${infer R}`
-  ? R
-  : never;
-
-// 文字列の末尾を削除する型
-export type PopString<S extends string> = S extends `${infer R}${string}`
-  ? R
-  : never;
 
 // 文字列をUnion型に変換する型
 export type StringToUnion<S extends string> = S extends `${infer F}${infer R}`
